% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.properties.R
\name{read.properties}
\alias{read.properties}
\title{Read a Java Properties File for Use in RSB Applications}
\usage{
read.properties(file, fields = NULL, encoding = "UTF-8")
}
\arguments{
\item{file}{properties file, either a character string (path) or a connection}

\item{fields}{subset of field names, if NULL, all fields are included}

\item{encoding}{encoding of the properties file to read (default value \code{"UTF-8"})}
}
\value{
list with key-value pairs
}
\description{
Read a Java Properties File for Use in RSB Applications
}
\examples{
tf <- system.file("testFiles", "test0.properties", package = "properties")
read.properties(tf, fields = c("key2", "key3"))
}
\seealso{
\code{\link{write.properties}} for writing properties files
}
\author{
Tobias Verbeke, Daan Seynaeve
}
